Console Shutdown Rework
Disconnected the console from the DiscordBot in a way. The shutdownBot method now only will kill the bot if it is running, not dispose of the console window. The console window must be manually closed in addition to shutting down the bot.

The reason for this is to support a user of Foundation's abililty to restart the bot if they so desire. Closing the console window will only terminate the app once the bot has been shutdown. If there is no support for a tray icon, then the console will not allow itself to be closed without first shutting down the bot.