SQLite Manager Pt. 1
So my friend introduced me to SQLite and I was like "yo! this looks like it would be a lot better than storing maps of maps of maps and then parsing them into .json objects!'.

So I started making some helpers to interact with an SQLite database for storing the configs and data of the commands. We'll see how this goes. It is currently incomplete and untested.