Menu Rework
Menus are now handled differently.

Before, Menus would require that a message had already been created and they would just handle the subsequent updates to that message.

Now, the Menus create their own message and handle it internally. Access to the Message object is limited to the update() method of the deriving class. Because this message is queued to prevent JDA from deadlocking, there technically is no garuntee that this message will complete immediately enough for any other actions to take place after its initialization. This should be kept in mind for its implementation, though I can't imagine (I risk saying this) that it would be a great impediment to functionality.