SubCommands Complete Implementation
Adding SubCommands was step one. Integrating them into the rest of the Foundation was step 2, which I believe to be complete now.

The command list *contains* all root commands, that is commands with no parents.

Each Command has a list that can contain subcommands.

The iterator returned by the commands list in Command.java will iterator over the command tree following a preorder traversal pattern.

Because of this iterator, all sub commands are saved and unloaded via the unloadAll and saveAll static mathod in Command.java. Therefore parent commands to not have to handle this manually.

The full name a command can be defined as its own name if it is a root command, or its parent's name, followed by a dot, and then its own name.

The Help command will now display subcommands of a command immediately after the parent command, if the member calling for the help menu has permissions to that subcommand. The order in which commands show in the help menu is the same preorder traversal that the iterator returns.