SubCommand Fix
Are you surprised?

I noticed that addCommand and getCommand were originally designed to just get root commands, but with the addition of the preorder iterator, they were now running over all commands and their sub commands looking for instances of the cmd name.

Instead of trying to fix this, I just changed the check be against the full name of the command. This does allow for sub commands to be run directly via 'bot_prefix parent_cmd.child_cmd', but I may add some logic later to set whether commands are allowed to be run directly or whether they have to be invoked via another command object.